% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Preparation of Raw Data}
\usage{
prepareData(rawData, blocksize = 100)
}
\arguments{
\item{rawData}{Raw data observations as depicted in the sampleData data structure}

\item{blocksize}{This input determines number of raw data observations per tabulated block}
}
\value{
From raw date, return objects to the global environment that can be used in
'removeAnomaly' and 'plotQQ' fucntions.
}
\description{
Takes raw data frame as an input and generates tabulated state vector and initial
data structures necessary for chi-square plot generation.  Functionality testign can
be conducted with the 'sampleData' Rda file included with this package.
}
