% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_myfirstApp.R
\name{run_myfirstApp}
\alias{run_myfirstApp}
\title{Runs a shiny app in my package}
\usage{
run_myfirstApp(dir, ...)
}
\arguments{
\item{dir}{Character string for directory in this package}

\item{...}{Additional options passed to shinyappDir}
}
\value{
A printed shiny app
}
\description{
Runs as shiny app
}
