% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCAC.R
\name{runMCAC}
\alias{runMCAC}
\title{Run MCAC Shiny App}
\usage{
runMCAC(dir, ...)
}
\arguments{
\item{dir}{Character string for directory in this package}

\item{...}{Additional options passed to shinyappDir}
}
\value{
A printed shiny app
}
\description{
Runs as shiny app
}
