% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Runs a shiny app in my package}
\usage{
prepareData(rawData, blocksize = 100)
}
\arguments{
\item{RawData}{Raw data observations as depicted in the SampleData data structure}

\item{Blocksize}{This input determines number of raw data observations per tabulated block}
}
\value{
This transforms raw data input into the tabulated state vector and returns input arguments needed for remove function
}
\description{
Takes raw data frame as an input and generates tabulated state vector and initial
data structures necessary for chi-square plot generation
}
